04-mapper-frameworks/src/test/java/net/franckbenault/microbench/mapper/MapperWithModelMapperOptimizedTest.java
package net.franckbenault.microbench.mapper;

import static org.junit.Assert.*;

import org.junit.Test;

import net.franckbenault.microbench.TestUtils;
import net.franckbenault.microbench.destination.OrderDTO;
import net.franckbenault.microbench.source.Order;

public class MapperWithModelMapperOptimizedTest {

	@Test
	public void testAsOrderDTO() {
		Order order = TestUtils.getOrder();
		MapperWithModelMapperOptimized mapper = new MapperWithModelMapperOptimized();
		OrderDTO orderDTO = mapper.asOrderDTO(order);
		
		assertNotNull(orderDTO);
		TestUtils.checkMapping(order, orderDTO);
	}
	
	@Test
	public void testAsOrderDTOEmpty() {
		Order order = new Order();
		MapperWithModelMapperOptimized mapper = new MapperWithModelMapperOptimized();
		OrderDTO orderDTO = mapper.asOrderDTO(order);
		
		assertNotNull(orderDTO);
		TestUtils.checkMapping(order, orderDTO);
	}
	
	@Test
	public void testAsOrder() {
		OrderDTO orderDTO = TestUtils.getOrderDTO();
		MapperWithModelMapperOptimized mapper = new MapperWithModelMapperOptimized();
		Order order = mapper.asOrder(orderDTO);
		
		assertNotNull(order);
		TestUtils.checkMapping(order, orderDTO);
	}
	
	@Test
	public void testAsOrderEmpty() {
		OrderDTO orderDTO = new OrderDTO();
		MapperWithModelMapperOptimized mapper = new MapperWithModelMapperOptimized();
		Order order = mapper.asOrder(orderDTO);
		
		assertNotNull(order);
		TestUtils.checkMapping(order, orderDTO);
	}

}
